requirements.txt
aioredis==1.3.1
amqp==5.0.7
asn1crypto==1.4.0
async-timeout==4.0.2
autopep8==1.5.6
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.3
dnspython==1.16.0
ecdsa==0.14.1
elasticsearch==7.16.1
et-xmlfile==1.1.0
eventlet==0.31.1
fastapi==0.63.0
fastapi-limiter==0.1.4
Flask==1.1.2
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==1.2
kombu==5.2.2
MarkupSafe==1.1.1
numpy==1.21.4
openpyxl==3.0.7
pandas==1.3.1
passlib==1.7.4
prompt-toolkit==3.0.24
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
pydub==0.25.1
pymongo==3.12.0
python-dateutil==2.8.1
python-http-client==3.3.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.3
redis==3.5.3
redisbeat==1.2.4
requests==2.25.1
rsa==4.7.2
sendgrid==6.8.1
sentry-sdk==1.0.0
six==1.15.0
starkbank-ecdsa==2.0.3
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
